requirements.txt
django.1.11
window OS